Move murmur3 into single directory
