+++ .gitignore
@@ -23,3 +23,4 @@ script/.*.swp
 .*.swp
 /test/s5_client
 src/b64/*.o
+/src/murmur3/murmur3.o
